Move logic back to MySpace
